LLM-projects/tutorials/LangChain-VectorDB-Learning/Projects/7. Making LLMs Interact with the World Using Tools/LangChain_Agents_and_Toolkits/requirements.txt
langchain==0.0.208 
deeplake 
openai 
google-search-results 
tiktoken
python-dotenv
google-api-python-client
wikipedia
Wolframalpha